# orders-platform/contracts/openapi.yaml
---
openapi: 3.1.0
components:
  schemas:
    ApiError:
      type: object
      properties:
        message:
          type: string
    CreateProductRequest:
      type: object
      properties:
        name:
          type: string
        sku:
          type: string
        quantity:
          type: integer
          format: int32
    Instant:
      type: string
      format: date-time
      examples:
      - 2022-03-10T16:15:50Z
    ProductResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        sku:
          type: string
        quantity:
          type: integer
          format: int32
    StockAdjustmentRequest:
      type: object
      properties:
        delta:
          type: integer
          format: int32
    StockMovementResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        productId:
          type: integer
          format: int64
        delta:
          type: integer
          format: int32
        previousQuantity:
          type: integer
          format: int32
        resultingQuantity:
          type: integer
          format: int32
        occurredAt:
          $ref: "#/components/schemas/Instant"
    UpdateProductRequest:
      type: object
      properties:
        name:
          type: string
paths:
  /inventory/products:
    get:
      responses:
        "200":
          description: Listado de productos
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProductResponse"
      summary: List
      tags:
      - Product Resource
    post:
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProductRequest"
        required: true
      responses:
        "201":
          description: Producto creado
          headers:
            Location:
              description: URI del producto creado
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductResponse"
        "400":
          description: "Datos inválidos (nombre o sku en blanco, cantidad negativa)"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "409":
          description: El SKU ya existe en el inventario
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
      summary: Create
      tags:
      - Product Resource
  /inventory/products/{id}:
    put:
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProductRequest"
        required: true
      responses:
        "200":
          description: Producto actualizado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductResponse"
        "400":
          description: "Datos inválidos (nombre en blanco, o un campo no admitido\
            \ como quantity)"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "404":
          description: No existe un producto con ese id
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
      summary: Update
      tags:
      - Product Resource
    get:
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: Producto encontrado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductResponse"
        "404":
          description: No existe un producto con ese id
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
      summary: Get By Id
      tags:
      - Product Resource
    delete:
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "204":
          description: Producto eliminado
        "404":
          description: No existe un producto con ese id
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
      summary: Delete
      tags:
      - Product Resource
  /inventory/products/{id}/stock-adjustments:
    post:
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StockAdjustmentRequest"
        required: true
      responses:
        "200":
          description: Stock ajustado; devuelve el producto con la cantidad resultante
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductResponse"
        "400":
          description: Ajuste inválido (delta cero o resultado fuera del rango admitido)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "404":
          description: No existe un producto con ese id
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "409":
          description: La salida dejaría el stock en negativo
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
      summary: Adjust Stock
      tags:
      - Product Resource
  /inventory/products/{id}/stock-movements:
    get:
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: "Historial de movimientos de stock, del más reciente al má\
            s antiguo"
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/StockMovementResponse"
        "404":
          description: No existe un producto con ese id
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
      summary: Stock Movements
      tags:
      - Product Resource
info:
  title: api API
  version: 1.0.0-SNAPSHOT